okay, tf did I do...
I somehow got the collisions to work properly, and I don't know how. if I change it in any sort of way it doesn't seem to work. and the best part is, for some reason I got it working, but BACKWARDS... I will just change the chances and image that is displayed to fix it, but I don't know how that happened.